Finding overlapping genomic tuples

Various methods for finding/counting overlaps between objects containing genomic tuples. This man page describes the methods that operate on GTuples and GTuplesList objects.

NOTE: The findOverlaps generic function is defined and documented in the IRanges package. The findOverlaps method for GenomicRanges and GRangesList objects are defined and documented in the GenoicRanges package.

GTuples and GTuplesList objects also support countOverlaps, overlapsAny, and subsetByOverlaps thanks to the default methods defined in the IRanges package and to the findOverlaps and countOverlaps methods defined in this package and documented below.

# S4 method for GTuples,GTuples
findOverlaps(query, subject,
    maxgap = -1L, minoverlap = 0L,
    type = c("any", "start", "end", "within", "equal"),
    select = c("all", "first", "last", "arbitrary"),
    ignore.strand = FALSE)

# S4 method for GTuples,GTuples
countOverlaps(query, subject,
    maxgap = -1L, minoverlap = 0L,
    type = c("any", "start", "end", "within", "equal"),
    ignore.strand = FALSE)

Arguments

query, subject
A GTuples or GTuplesList object.

type
See details below.

maxgap, minoverlap
See ?findOverlaps in the IRanges package for a description of these arguments.

select
When select is "all" (the default), the results are returned as a Hits object. Otherwise the returned value is an integer vector parallel to query (i.e. same length) containing the first, last, or arbitrary overlapping interval in subject, with NA indicating intervals that did not overlap any intervals in subject.

ignore.strand
When set to TRUE, the strand information is ignored in the overlap calculations.

Details

The findOverlaps-based methods involving genomic tuples, either through GTuples or GTuplesList objects, can search for tuple-tuple, tuple-range and range-tuple overlaps. Each of these are described below, with attention paid to the important special case of finding "equal tuple-tuple overlaps".

Equal tuple-tuple overlaps

When the query and the subject are both GTuples objects and type = "equal", findOverlaps uses the seqnames (seqnames), positions (tuples,GTuples-method) and strand (strand) to determine which tuples from the query exactly match those in the subject, where a strand value of "*" is treated as occurring on both the "+" and "-" strand. An overlap is recorded when a tuple in the query and a tuple in the subject have the same sequence name, have a compatible pairing of strands (e.g. "+"/"+", "-"/"-", "*"/"+", "*"/"-", etc.), and have identical positions. NOTE: Equal tuple-tuple overlaps can only be computed if size(query) is equal to size(subject).

Other tuple-tuple overlaps

When the query and the subject are GTuples or GTuplesList objects and type = "any", "start", "end" or "within", findOverlaps treats the tuples as if they were ranges, with ranges given by \([pos_{1}, pos_{m}]\) and where \(m\) is the size,GTuples-method of the tuples. This is done via inheritance so that a GTuples (resp. GTuplesList) object is treated as a GRanges (resp. GRangesList) and the appropriate findOverlaps method is dispatched upon. NOTE: This is the only type of overlap finding available when either the query and subject are GTuplesList objects. This is following the behaviour of findOverlaps,GRangesList,GRangesList-method that allows type = "any", "start", "end" or "within" but does not allow type = "equal".

tuple-range and range-tuple overlaps

When one of the query and the subject is not a GTuples or GTuplesList objects, findOverlaps treats the tuples as if they were ranges, with ranges given by \([pos_{1}, pos_{m}]\) and where \(m\) is the size,GTuples-method of the tuples. This is done via inheritance so that a GTuples (resp. GTuplesList) object is treated as a GRanges (resp. GRangesList) and the appropriate findOverlaps method is dispatched upon.

In the context of findOverlaps, a feature is a collection of tuples/ranges that are treated as a single entity. For GTuples objects, a feature is a single tuple; while for GTuplesList objects, a feature is a list element containing a set of tuples. In the results, the features are referred to by number, which run from 1 to length(query)/length(subject).

Value

For findOverlaps, either a Hits object when select = "all" or an integer vector otherwise.

For countOverlaps, an integer vector containing the tabulated query overlap hits.

For overlapsAny a logical vector of length equal to the number of tuples/ranges in query indicating those that overlap any of the tuples/ranges in subject.

For subsetByOverlaps an object of the same class as query containing the subset that overlapped at least one entity in subject.

See also

Please see the package vignette for an extended discussion of overlaps involving genomic tuples, which is available by typing vignette(topic = 'GenomicTuplesIntroduction', package = 'GenomicTuples') at the R prompt.

findOverlaps

findOverlaps

Hits-class

GTuples-class

GTuplesList-class

GRanges-class

GRangesList-class

Examples

## GTuples object containing 3-tuples:
gt3 <- GTuples(seqnames = c('chr1', 'chr1', 'chr1', 'chr1', 'chr2'),
               tuples = matrix(c(10L, 10L, 10L, 10L, 10L, 20L, 20L, 20L, 25L,
                                 20L, 30L, 30L, 35L, 30L, 30L), ncol = 3),
               strand = c('+', '-', '*', '+', '+'))

## GTuplesList object
gtl3 <- GTuplesList(A = gt3[1:3], B = gt3[4:5])

## Find equal genomic tuples:
findOverlaps(gt3, gt3, type = 'equal')

#> Hits object with 5 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   [2]         2           2
#>   [3]         3           3
#>   [4]         4           4
#>   [5]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

## Note that this is different to the results if the tuples are treated as 
## ranges since this ignores the "internal positions" (pos2):
findOverlaps(granges(gt3), granges(gt3), type = 'equal')

#> Hits object with 7 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   [2]         1           4
#>   [3]         2           2
#>   [4]         3           3
#>   [5]         4           1
#>   [6]         4           4
#>   [7]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

## Scenarios where tuples are treated as ranges:
findOverlaps(gt3, gt3, type = 'any')

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Note: method with signature ‘GRanges#GRanges’ chosen for function ‘findOverlaps’,
#>  target signature ‘GTuples#GTuples’.
#>  "GenomicRanges#GenomicRanges" would also be valid

#> Hits object with 13 hits and 0 metadata columns:
#>        queryHits subjectHits
#>        <integer>   <integer>
#>    [1]         1           3
#>    [2]         1           1
#>    [3]         1           4
#>    [4]         2           3
#>    [5]         2           2
#>    ...       ...         ...
#>    [9]         3           4
#>   [10]         4           3
#>   [11]         4           1
#>   [12]         4           4
#>   [13]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

findOverlaps(gt3, gt3, type = 'start')

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Hits object with 13 hits and 0 metadata columns:
#>        queryHits subjectHits
#>        <integer>   <integer>
#>    [1]         1           3
#>    [2]         1           1
#>    [3]         1           4
#>    [4]         2           3
#>    [5]         2           2
#>    ...       ...         ...
#>    [9]         3           4
#>   [10]         4           3
#>   [11]         4           1
#>   [12]         4           4
#>   [13]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

findOverlaps(gt3, gt3, type = 'end')

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Hits object with 7 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   [2]         1           4
#>   [3]         2           2
#>   [4]         3           3
#>   [5]         4           1
#>   [6]         4           4
#>   [7]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

findOverlaps(gt3, gt3, type = 'within')

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Hits object with 10 hits and 0 metadata columns:
#>        queryHits subjectHits
#>        <integer>   <integer>
#>    [1]         1           3
#>    [2]         1           1
#>    [3]         1           4
#>    [4]         2           3
#>    [5]         2           2
#>    [6]         3           3
#>    [7]         4           3
#>    [8]         4           1
#>    [9]         4           4
#>   [10]         5           5
#>   -------
#>   queryLength: 5 / subjectLength: 5

## Overlapping a GTuples and a GTuplesList object (tuples treated as ranges):
table(!is.na(findOverlaps(gtl3, gt3, select="first")))

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> 
#> TRUE 
#>    2 

countOverlaps(gtl3, gt3)

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> A B 
#> 4 4 

findOverlaps(gtl3, gt3)

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Hits object with 8 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           3
#>   [2]         1           1
#>   [3]         1           4
#>   [4]         1           2
#>   [5]         2           3
#>   [6]         2           1
#>   [7]         2           4
#>   [8]         2           5
#>   -------
#>   queryLength: 2 / subjectLength: 5

subsetByOverlaps(gtl3, gt3)

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> GTuplesList object of length 2:
#> $A 
#> GTuples object with 3 x 3-tuples and 0 metadata columns:
#>       seqnames pos1 pos2 pos3 strand
#>   [1]     chr1   10   20   30      +
#>   [2]     chr1   10   20   30      -
#>   [3]     chr1   10   20   35      *
#> 
#> $B 
#> GTuples object with 2 x 3-tuples and 0 metadata columns:
#>       seqnames pos1 pos2 pos3 strand
#>   [1]     chr1   10   25   30      +
#>   [2]     chr2   10   20   30      +
#> 
#> -------
#> seqinfo: 2 sequences from an unspecified genome; no seqlengths

countOverlaps(gtl3, gt3, type = "start")

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> A B 
#> 4 4 

findOverlaps(gtl3, gt3, type = "start")

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> Hits object with 8 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           3
#>   [2]         1           1
#>   [3]         1           4
#>   [4]         1           2
#>   [5]         2           3
#>   [6]         2           1
#>   [7]         2           4
#>   [8]         2           5
#>   -------
#>   queryLength: 2 / subjectLength: 5

subsetByOverlaps(gtl3, gt3, type = "start")

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> GTuplesList object of length 2:
#> $A 
#> GTuples object with 3 x 3-tuples and 0 metadata columns:
#>       seqnames pos1 pos2 pos3 strand
#>   [1]     chr1   10   20   30      +
#>   [2]     chr1   10   20   30      -
#>   [3]     chr1   10   20   35      *
#> 
#> $B 
#> GTuples object with 2 x 3-tuples and 0 metadata columns:
#>       seqnames pos1 pos2 pos3 strand
#>   [1]     chr1   10   25   30      +
#>   [2]     chr2   10   20   30      +
#> 
#> -------
#> seqinfo: 2 sequences from an unspecified genome; no seqlengths

findOverlaps(gtl3, gt3, select = "first")

#> Warning: 'type' is not 'equal' so coercing 'query' and 'subject' to 'GRanges' objects (see docs for details)

#> [1] 1 1